finished hw05
