+++ controllers/auth.js
@@ -75,16 +75,20 @@ const updateSubscription = async (req, res) => {
 const updateAvatar = async (req, res) => {
 	const { _id } = req.user;
 	const { path: tempUpload, originalname } = req.file;
-	await jimpResizer(tempUpload);
-	const newName = `${_id}_${originalname}`;
-	const distUpload = path.join(avatarsDir, newName);
+	try {
+		await jimpResizer(tempUpload);
+		const newName = `${_id}_${originalname}`;
+		const distUpload = path.join(avatarsDir, newName);
 
-	await fs.rename(tempUpload, distUpload);
-	const avatarURL = path.join("avatars", newName);
-	await User.findByIdAndUpdate(_id, { avatarURL });
-	res.json({
-		avatarURL,
-	});
+		await fs.rename(tempUpload, distUpload);
+		const avatarURL = path.join("avatars", newName);
+		await User.findByIdAndUpdate(_id, { avatarURL });
+		res.json({
+			avatarURL,
+		});
+	} catch (error) {
+		await fs.unlink(tempUpload);
+	}
 };
 
 module.exports = {
