+++ package.json
@@ -25,8 +25,7 @@
 		"mongoose": "^7.0.3",
 		"morgan": "1.10.0",
 		"multer": "^1.4.5-lts.1",
-		"nanoid": "^3.3.6",
-		"supertest": "^6.3.3"
+		"nanoid": "^3.3.6"
 	},
 	"devDependencies": {
 		"eslint": "7.19.0",
@@ -36,6 +35,7 @@
 		"eslint-plugin-node": "^11.1.0",
 		"eslint-plugin-promise": "5.2.0",
 		"jest": "^29.5.0",
-		"nodemon": "2.0.15"
+		"nodemon": "2.0.15",
+		"supertest": "^6.3.3"
 	}
 }
